requirements.txt
feedparser
PyYAML
click